==> Revise_Experiment/Revisi 2/1a.DataPreProcessing_Sensor1.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
%matplotlib inline

# tabel setting
# pd.set_option('display.max_columns',None)
# pd.set_option('display.max_rows',None)
# pd.set_option('display.precision', 2)
# pd.options.display.float_format = '{:,.2f}'.format

# Import Dataset

dataset_dir = './Dataset/NCC-2 Dataset Simultaneous Botnet Dataset/'
sensor1_path = dataset_dir + 'sensor1/'
sensor1_path = dataset_dir + 'sensor1/'
sensor1_path = dataset_dir + 'sensor1/'
allsensor_path = dataset_dir + 'all-sensors/'

os.listdir(sensor1_path)

# Sensor 1 .binetflow

# df1 -> all, df2 -> botnet only, df3 -> normal only
sensor1 = pd.read_csv(sensor1_path + 'sensor1.binetflow')

sensor1.shape

sensor1.columns

sensor1.dtypes

sensor1.head()

sensor1_cat = sensor1.select_dtypes('object')
sensor1_num = sensor1.select_dtypes(['int64', 'float64'])

print('Kolom Categorical :', sensor1_cat.columns.value_counts().count())
print('Kolom Numerical :', sensor1_num.columns.value_counts().count())

del sensor1_cat, sensor1_num

## Check Missing Value

# Cek Missing Data
total_missing = sensor1.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1.isnull().sum()/sensor1.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

## Cek Kolom

### Start Time

pd.DataFrame(sensor1['StartTime'].value_counts()).reset_index()

### Dir

pd.DataFrame(sensor1['Dir'].value_counts()).reset_index()

### State

pd.DataFrame(sensor1['State'].value_counts()).reset_index()

### BotNet Name

pd.DataFrame(sensor1['BotnetName'].value_counts()).reset_index()

### SrcAddr

pd.DataFrame(sensor1['SrcAddr'].value_counts()).reset_index()

### DstAddr

pd.DataFrame(sensor1['DstAddr'].value_counts()).reset_index()

### Proto

pd.DataFrame(sensor1['Proto'].value_counts()).reset_index()

### Sport

pd.DataFrame(sensor1['Sport'].value_counts()).reset_index()

### Dport

pd.DataFrame(sensor1['Dport'].value_counts()).reset_index()

### dTos

pd.DataFrame(sensor1['dTos'].value_counts()).reset_index()

### sTos

pd.DataFrame(sensor1['sTos'].value_counts()).reset_index()

## Labeling

pd.DataFrame(sensor1['Label'].value_counts()).reset_index()

def categorize_label(label):
    label = label.lower()  
    if 'botnet' in label and 'spam' in label:
        return 2
    elif 'botnet' in label:
        return 1
    else:
        return 0

sensor1['Label'] = sensor1['Label'].apply(categorize_label)

sensor1.head()

sensor1['Label'].value_counts()

## Cleaning Data

### Drop Duplicates

sensor1 = sensor1.drop_duplicates()

sensor1.shape

### Drop dTos

cek = sensor1[(sensor1['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor1 = sensor1.drop(cek.index)

# Cek Missing Data
total_missing = sensor1.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1.isnull().sum()/sensor1.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop sTos

cek = sensor1[(sensor1['sTos'].isnull())]

cek['Label'].value_counts()

sensor1 = sensor1.drop(cek.index)

# Cek Missing Data
total_missing = sensor1.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1.isnull().sum()/sensor1.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop Dport

cek = sensor1[sensor1['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor1 = sensor1.drop(cek.index)

# Cek Missing Data
total_missing = sensor1.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1.isnull().sum()/sensor1.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop Sport

cek = sensor1[sensor1['Sport'].isnull()]

cek['Label'].value_counts()

# cek = cek[cek['Label'] == 0]

# cek['Label'].value_counts()

# sensor1 = sensor1.drop(cek.index)

# Cek Missing Data
total_missing = sensor1.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1.isnull().sum()/sensor1.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop State

cek = sensor1[sensor1['State'].isnull()]

cek['Label'].value_counts()

# Cek Missing Data
total_missing = sensor1.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1.isnull().sum()/sensor1.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

## Imputing

### dTos

sensor1['dTos'].value_counts()

cek = sensor1[sensor1['dTos'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value1 = sensor1[sensor1['Label'] == 1]['dTos'].mode()[0]
mode_value2 = sensor1[sensor1['Label'] == 2]['dTos'].mode()[0]

sensor1.loc[(sensor1['dTos'].isnull()) & (sensor1['Label'] == 1), 'dTos'] = mode_value1
sensor1.loc[(sensor1['dTos'].isnull()) & (sensor1['Label'] == 2), 'dTos'] = mode_value2

# Cek Missing Data
total_missing = sensor1.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1.isnull().sum()/sensor1.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Dport

sensor1['Dport'].value_counts()

cek = sensor1[sensor1['Dport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor1[sensor1['Label'] == 1]['Dport'].mode()[0]
sensor1['Dport'] = sensor1['Dport'].fillna(mode_value)

### Sport

sensor1['Sport'].value_counts()

cek = sensor1[sensor1['Sport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor1[sensor1['Label'] == 1]['Sport'].mode()[0]
sensor1['Sport'] = sensor1['Sport'].fillna(mode_value)

### State

sensor1['State'].value_counts()

cek = sensor1[sensor1['State'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor1[sensor1['Label'] == 1]['State'].mode()[0]
sensor1['State'] = sensor1['State'].fillna(mode_value)

# Cek Missing Data
total_missing = sensor1.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1.isnull().sum()/sensor1.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# sensor1.to_csv('./Dataset/sensor1_clean.csv', index=False)

## Categorical Encoding

# df.to_csv('./Dataset/df_clean.csv')

dataset = pd.DataFrame(sensor1[[
    'StartTime',
    'SrcAddr',
    'DstAddr',
    'Sport',
    'Dport',
    'sTos',
    'dTos',
    'Dir',
    'State',
    'Proto',
    'Dur',
    'TotPkts',
    'TotBytes',
    'SrcBytes',
    'BotnetName',
    'Label',
    ]])

### Start Time

dataset['StartTime'] = pd.to_datetime(dataset['StartTime'], errors='coerce')

dataset['StartTimeHour'] = dataset['StartTime'].dt.hour
dataset['StartTimeMinute'] = dataset['StartTime'].dt.minute
dataset['StartTimeSecond'] = dataset['StartTime'].dt.second
dataset = dataset.drop(columns=['StartTime'])

### SrcAddr & DstAddr & Sport & Dport

# Frequency Encoding
def frequency_encoding(df):
    freq_map = df.value_counts().to_dict()  # Create frequency map
    df = df.map(freq_map)  # Apply frequency encoding
    return df

dataset['SrcAddr'] = frequency_encoding(dataset['SrcAddr'])
dataset['DstAddr'] = frequency_encoding(dataset['DstAddr'])
dataset['Sport'] = frequency_encoding(dataset['Sport'])
dataset['Dport'] = frequency_encoding(dataset['Dport'])

### Dir & State & dTos & sTos & Proto

# Label Encoding
from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()
dataset['Dir'] = le.fit_transform(dataset['Dir'].astype(str))
dataset['State'] = le.fit_transform(dataset['State'].astype(str))
dataset['sTos'] = le.fit_transform(dataset['sTos'].astype(str))
dataset['dTos'] = le.fit_transform(dataset['dTos'].astype(str))
dataset['BotnetName'] = le.fit_transform(dataset['BotnetName'].astype(str))
dataset['Proto'] = le.fit_transform(dataset['Proto'].astype(str))

dataset.shape

dataset.head()

dataset['isBotnet'] = dataset['Label'].apply(lambda x: 1 if x == 1 or x == 2 else 0)
dataset['isSpam'] = dataset['Label'].apply(lambda x: 1 if x == 2 else 0)

dataset.columns

# dataset.to_csv('./Dataset/sensor1_clean.csv', index=False)
