==> Revise_Experiment/Revisi 2/1b.DataPreProcessing_Sensor2.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
%matplotlib inline

# tabel setting
# pd.set_option('display.max_columns',None)
# pd.set_option('display.max_rows',None)
# pd.set_option('display.precision', 2)
# pd.options.display.float_format = '{:,.2f}'.format

# Import Dataset

dataset_dir = './Dataset/NCC-2 Dataset Simultaneous Botnet Dataset/'
sensor1_path = dataset_dir + 'sensor1/'
sensor2_path = dataset_dir + 'sensor2/'
sensor2_path = dataset_dir + 'sensor2/'
allsensor_path = dataset_dir + 'all-sensors/'

os.listdir(sensor2_path)

# Sensor 2 .binetflow

# df1 -> all, df2 -> botnet only, df3 -> normal only
sensor2 = pd.read_csv(sensor2_path + 'sensor2.binetflow')

sensor2.shape

sensor2.columns

sensor2.dtypes

sensor2.head()

sensor2_cat = sensor2.select_dtypes('object')
sensor2_num = sensor2.select_dtypes(['int64', 'float64'])

print('Kolom Categorical :', sensor2_cat.columns.value_counts().count())
print('Kolom Numerical :', sensor2_num.columns.value_counts().count())

del sensor2_cat, sensor2_num

## Check Missing Value

# Cek Missing Data
total_missing = sensor2.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2.isnull().sum()/sensor2.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

## Cek Kolom

### Start Time

pd.DataFrame(sensor2['StartTime'].value_counts()).reset_index()

### Dir

pd.DataFrame(sensor2['Dir'].value_counts()).reset_index()

### State

pd.DataFrame(sensor2['State'].value_counts()).reset_index()

### BotNet Name

pd.DataFrame(sensor2['BotnetName'].value_counts()).reset_index()

### SrcAddr

pd.DataFrame(sensor2['SrcAddr'].value_counts()).reset_index()

### DstAddr

pd.DataFrame(sensor2['DstAddr'].value_counts()).reset_index()

### Proto

pd.DataFrame(sensor2['Proto'].value_counts()).reset_index()

### Sport

pd.DataFrame(sensor2['Sport'].value_counts()).reset_index()

### Dport

pd.DataFrame(sensor2['Dport'].value_counts()).reset_index()

### dTos

pd.DataFrame(sensor2['dTos'].value_counts()).reset_index()

### sTos

pd.DataFrame(sensor2['sTos'].value_counts()).reset_index()

## Labeling

pd.DataFrame(sensor2['Label'].value_counts()).reset_index()

def categorize_label(label):
    label = label.lower()  
    if 'botnet' in label and 'spam' in label:
        return 2
    elif 'botnet' in label:
        return 1
    else:
        return 0

sensor2['Label'] = sensor2['Label'].apply(categorize_label)

sensor2.head()

sensor2['Label'].value_counts()

## Cleaning Data

### Drop Duplicates

sensor2 = sensor2.drop_duplicates()

sensor2.shape

### Drop dTos

cek = sensor2[(sensor2['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor2 = sensor2.drop(cek.index)

# Cek Missing Data
total_missing = sensor2.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2.isnull().sum()/sensor2.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop sTos

cek = sensor2[(sensor2['sTos'].isnull())]

cek['Label'].value_counts()

sensor2 = sensor2.drop(cek.index)

# Cek Missing Data
total_missing = sensor2.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2.isnull().sum()/sensor2.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop Dport

cek = sensor2[sensor2['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor2 = sensor2.drop(cek.index)

# Cek Missing Data
total_missing = sensor2.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2.isnull().sum()/sensor2.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop Sport

cek = sensor2[sensor2['Sport'].isnull()]

cek['Label'].value_counts()

# cek = cek[cek['Label'] == 0]

# cek['Label'].value_counts()

# sensor2 = sensor2.drop(cek.index)

# Cek Missing Data
total_missing = sensor2.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2.isnull().sum()/sensor2.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop State

cek = sensor2[sensor2['State'].isnull()]

cek['Label'].value_counts()

# Cek Missing Data
total_missing = sensor2.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2.isnull().sum()/sensor2.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

## Imputing

### dTos

sensor2['dTos'].value_counts()

cek = sensor2[sensor2['dTos'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value1 = sensor2[sensor2['Label'] == 1]['dTos'].mode()[0]
mode_value2 = sensor2[sensor2['Label'] == 2]['dTos'].mode()[0]

sensor2.loc[(sensor2['dTos'].isnull()) & (sensor2['Label'] == 1), 'dTos'] = mode_value1
sensor2.loc[(sensor2['dTos'].isnull()) & (sensor2['Label'] == 2), 'dTos'] = mode_value2

# Cek Missing Data
total_missing = sensor2.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2.isnull().sum()/sensor2.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Dport

sensor2['Dport'].value_counts()

cek = sensor2[sensor2['Dport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor2[sensor2['Label'] == 1]['Dport'].mode()[0]
sensor2['Dport'] = sensor2['Dport'].fillna(mode_value)

### Sport

sensor2['Sport'].value_counts()

cek = sensor2[sensor2['Sport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor2[sensor2['Label'] == 1]['Sport'].mode()[0]
sensor2['Sport'] = sensor2['Sport'].fillna(mode_value)

### State

sensor2['State'].value_counts()

cek = sensor2[sensor2['State'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor2[sensor2['Label'] == 1]['State'].mode()[0]
sensor2['State'] = sensor2['State'].fillna(mode_value)

# Cek Missing Data
total_missing = sensor2.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2.isnull().sum()/sensor2.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# sensor2.to_csv('./Dataset/sensor2_clean.csv', index=False)

## Categorical Encoding

# df.to_csv('./Dataset/df_clean.csv')

dataset = pd.DataFrame(sensor2[[
    'StartTime',
    'SrcAddr',
    'DstAddr',
    'Sport',
    'Dport',
    'sTos',
    'dTos',
    'Dir',
    'State',
    'Proto',
    'Dur',
    'TotPkts',
    'TotBytes',
    'SrcBytes',
    'BotnetName',
    'Label',
    ]])

### Start Time

dataset['StartTime'] = pd.to_datetime(dataset['StartTime'], errors='coerce')

dataset['StartTimeHour'] = dataset['StartTime'].dt.hour
dataset['StartTimeMinute'] = dataset['StartTime'].dt.minute
dataset['StartTimeSecond'] = dataset['StartTime'].dt.second
dataset = dataset.drop(columns=['StartTime'])

### SrcAddr & DstAddr & Sport & Dport

# Frequency Encoding
def frequency_encoding(df):
    freq_map = df.value_counts().to_dict()  # Create frequency map
    df = df.map(freq_map)  # Apply frequency encoding
    return df

dataset['SrcAddr'] = frequency_encoding(dataset['SrcAddr'])
dataset['DstAddr'] = frequency_encoding(dataset['DstAddr'])
dataset['Sport'] = frequency_encoding(dataset['Sport'])
dataset['Dport'] = frequency_encoding(dataset['Dport'])

### Dir & State & dTos & sTos & Proto

# Label Encoding
from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()
dataset['Dir'] = le.fit_transform(dataset['Dir'].astype(str))
dataset['State'] = le.fit_transform(dataset['State'].astype(str))
dataset['sTos'] = le.fit_transform(dataset['sTos'].astype(str))
dataset['dTos'] = le.fit_transform(dataset['dTos'].astype(str))
dataset['BotnetName'] = le.fit_transform(dataset['BotnetName'].astype(str))
dataset['Proto'] = le.fit_transform(dataset['Proto'].astype(str))

dataset.shape

dataset.head()

dataset['isBotnet'] = dataset['Label'].apply(lambda x: 1 if x == 1 or x == 2 else 0)
dataset['isSpam'] = dataset['Label'].apply(lambda x: 1 if x == 2 else 0)

dataset.columns

# dataset.to_csv('./Dataset/sensor2_clean.csv', index=False)
