==> Revise_Experiment/Revisi 2/1c.DataPreProcessing_Sensor3.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
%matplotlib inline

# tabel setting
# pd.set_option('display.max_columns',None)
# pd.set_option('display.max_rows',None)
# pd.set_option('display.precision', 2)
# pd.options.display.float_format = '{:,.2f}'.format

# Import Dataset

dataset_dir = './Dataset/NCC-2 Dataset Simultaneous Botnet Dataset/'
sensor1_path = dataset_dir + 'sensor1/'
sensor2_path = dataset_dir + 'sensor2/'
sensor3_path = dataset_dir + 'sensor3/'
allsensor_path = dataset_dir + 'all-sensors/'

os.listdir(sensor3_path)

# Sensor 3 .binetflow

# df1 -> all, df2 -> botnet only, df3 -> normal only
sensor3 = pd.read_csv(sensor3_path + 'sensor3.binetflow')

sensor3.shape

sensor3.columns

sensor3.dtypes

sensor3.head()

sensor3_cat = sensor3.select_dtypes('object')
sensor3_num = sensor3.select_dtypes(['int64', 'float64'])

print('Kolom Categorical :', sensor3_cat.columns.value_counts().count())
print('Kolom Numerical :', sensor3_num.columns.value_counts().count())

del sensor3_cat, sensor3_num

## Check Missing Value

# Cek Missing Data
total_missing = sensor3.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3.isnull().sum()/sensor3.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

## Cek Kolom

### Start Time

pd.DataFrame(sensor3['StartTime'].value_counts()).reset_index()

### Dir

pd.DataFrame(sensor3['Dir'].value_counts()).reset_index()

### State

pd.DataFrame(sensor3['State'].value_counts()).reset_index()

### BotNet Name

pd.DataFrame(sensor3['BotnetName'].value_counts()).reset_index()

### SrcAddr

pd.DataFrame(sensor3['SrcAddr'].value_counts()).reset_index()

### DstAddr

pd.DataFrame(sensor3['DstAddr'].value_counts()).reset_index()

### Proto

pd.DataFrame(sensor3['Proto'].value_counts()).reset_index()

### Sport

pd.DataFrame(sensor3['Sport'].value_counts()).reset_index()

### Dport

pd.DataFrame(sensor3['Dport'].value_counts()).reset_index()

### dTos

pd.DataFrame(sensor3['dTos'].value_counts()).reset_index()

### sTos

pd.DataFrame(sensor3['sTos'].value_counts()).reset_index()

## Labeling

pd.DataFrame(sensor3['Label'].value_counts()).reset_index()

def categorize_label(label):
    label = label.lower()  
    if 'botnet' in label and 'spam' in label:
        return 2
    elif 'botnet' in label:
        return 1
    else:
        return 0

sensor3['Label'] = sensor3['Label'].apply(categorize_label)

sensor3.head()

sensor3['Label'].value_counts()

## Cleaning Data

### Drop Duplicates

sensor3 = sensor3.drop_duplicates()

sensor3.shape

### Drop dTos

cek = sensor3[(sensor3['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3 = sensor3.drop(cek.index)

# Cek Missing Data
total_missing = sensor3.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3.isnull().sum()/sensor3.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop sTos

cek = sensor3[(sensor3['sTos'].isnull())]

cek['Label'].value_counts()

sensor3 = sensor3.drop(cek.index)

# Cek Missing Data
total_missing = sensor3.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3.isnull().sum()/sensor3.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop Dport

cek = sensor3[sensor3['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3 = sensor3.drop(cek.index)

# Cek Missing Data
total_missing = sensor3.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3.isnull().sum()/sensor3.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop Sport

cek = sensor3[sensor3['Sport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3 = sensor3.drop(cek.index)

# Cek Missing Data
total_missing = sensor3.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3.isnull().sum()/sensor3.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Drop State

cek = sensor3[sensor3['State'].isnull()]

cek['Label'].value_counts()

# Cek Missing Data
total_missing = sensor3.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3.isnull().sum()/sensor3.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

## Imputing

### dTos

sensor3['dTos'].value_counts()

cek = sensor3[sensor3['dTos'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value1 = sensor3[sensor3['Label'] == 1]['dTos'].mode()[0]
mode_value2 = sensor3[sensor3['Label'] == 2]['dTos'].mode()[0]

sensor3.loc[(sensor3['dTos'].isnull()) & (sensor3['Label'] == 1), 'dTos'] = mode_value1
sensor3.loc[(sensor3['dTos'].isnull()) & (sensor3['Label'] == 2), 'dTos'] = mode_value2

# Cek Missing Data
total_missing = sensor3.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3.isnull().sum()/sensor3.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Dport

sensor3['Dport'].value_counts()

cek = sensor3[sensor3['Dport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor3[sensor3['Label'] == 1]['Dport'].mode()[0]
sensor3['Dport'] = sensor3['Dport'].fillna(mode_value)

### Sport

sensor3['Sport'].value_counts()

cek = sensor3[sensor3['Sport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor3[sensor3['Label'] == 1]['Sport'].mode()[0]
sensor3['Sport'] = sensor3['Sport'].fillna(mode_value)

### State

sensor3['State'].value_counts()

cek = sensor3[sensor3['State'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor3[sensor3['Label'] == 1]['State'].mode()[0]
sensor3['State'] = sensor3['State'].fillna(mode_value)

# Cek Missing Data
total_missing = sensor3.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3.isnull().sum()/sensor3.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# sensor3.to_csv('./Dataset/sensor3_clean.csv', index=False)

## Categorical Encoding

# df.to_csv('./Dataset/df_clean.csv')

dataset = pd.DataFrame(sensor3[[
    'StartTime',
    'SrcAddr',
    'DstAddr',
    'Sport',
    'Dport',
    'sTos',
    'dTos',
    'Dir',
    'State',
    'Proto',
    'Dur',
    'TotPkts',
    'TotBytes',
    'SrcBytes',
    'BotnetName',
    'Label',
    ]])

### Start Time

dataset['StartTime'] = pd.to_datetime(dataset['StartTime'], errors='coerce')

dataset['StartTimeHour'] = dataset['StartTime'].dt.hour
dataset['StartTimeMinute'] = dataset['StartTime'].dt.minute
dataset['StartTimeSecond'] = dataset['StartTime'].dt.second
dataset = dataset.drop(columns=['StartTime'])

### SrcAddr & DstAddr & Sport & Dport

# Frequency Encoding
def frequency_encoding(df):
    freq_map = df.value_counts().to_dict()  # Create frequency map
    df = df.map(freq_map)  # Apply frequency encoding
    return df

dataset['SrcAddr'] = frequency_encoding(dataset['SrcAddr'])
dataset['DstAddr'] = frequency_encoding(dataset['DstAddr'])
dataset['Sport'] = frequency_encoding(dataset['Sport'])
dataset['Dport'] = frequency_encoding(dataset['Dport'])

### Dir & State & dTos & sTos & Proto

# Label Encoding
from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()
dataset['Dir'] = le.fit_transform(dataset['Dir'].astype(str))
dataset['State'] = le.fit_transform(dataset['State'].astype(str))
dataset['sTos'] = le.fit_transform(dataset['sTos'].astype(str))
dataset['dTos'] = le.fit_transform(dataset['dTos'].astype(str))
dataset['BotnetName'] = le.fit_transform(dataset['BotnetName'].astype(str))
dataset['Proto'] = le.fit_transform(dataset['Proto'].astype(str))

dataset.shape

dataset.head()

dataset['isBotnet'] = dataset['Label'].apply(lambda x: 1 if x == 1 or x == 2 else 0)
dataset['isSpam'] = dataset['Label'].apply(lambda x: 1 if x == 2 else 0)

dataset.columns

# dataset.to_csv('./Dataset/sensor3_clean.csv', index=False)
