==> Revise_Experiment/Revisi 2/2.EncodeData.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
from sklearn.model_selection import train_test_split
%matplotlib inline

# tabel setting
pd.set_option('display.max_columns',None)
pd.set_option('display.max_rows',None)
pd.set_option('display.precision', 2)
pd.options.display.float_format = '{:,.2f}'.format

dataset_dir = '../Dataset/'

os.listdir(dataset_dir)

# Encode Categorical

# Frequency Encoding
def frequency_encoding(df):
    freq_map = df.value_counts().to_dict()  # Create frequency map
    df = df.map(freq_map)  # Apply frequency encoding
    return df

from sklearn.preprocessing import LabelEncoder
def encode(df):
    df = df.drop_duplicates()

    # Start Time
    df['StartTime'] = pd.to_datetime(df['StartTime'], errors='coerce')
    df['StartTimeHour'] = df['StartTime'].dt.hour
    df['StartTimeMinute'] = df['StartTime'].dt.minute
    df['StartTimeSecond'] = df['StartTime'].dt.second
    df = df.drop(columns=['StartTime'])

    # SrcAddr & DstAddr & Sport & Dport Freq Encoding
    df['SrcAddr'] = frequency_encoding(df['SrcAddr'])
    df['DstAddr'] = frequency_encoding(df['DstAddr'])
    df['Sport'] = frequency_encoding(df['Sport'])
    df['Dport'] = frequency_encoding(df['Dport'])

    # Dir & State & dTos & sTos & Proto Label Encoding
    le = LabelEncoder()
    df['Dir'] = le.fit_transform(df['Dir'].astype(str))
    df['State'] = le.fit_transform(df['State'].astype(str))
    df['sTos'] = le.fit_transform(df['sTos'].astype(str))
    df['dTos'] = le.fit_transform(df['dTos'].astype(str))
    df['BotnetName'] = le.fit_transform(df['BotnetName'].astype(str))
    df['Proto'] = le.fit_transform(df['Proto'].astype(str))

    return df

def add_target(df):
    df['isBotnet'] = df['Label'].apply(lambda x: 1 if x == 1 or x == 2 else 0)
    df['isSpam'] = df['Label'].apply(lambda x: 1 if x == 2 else 0)
    return df

# Sensor 1

df_1 = pd.read_csv(dataset_dir + 'sensor1_clean.csv')

df_1.shape

df_1.head()

df_1 = encode(df_1)
df_1 = add_target(df_1)

df_1.shape

df_1.head()

df_1 = df_1.drop(columns=['ActivityLabel','BotnetName','SensorId'])

df_1.to_csv(dataset_dir + 'sensor1_encoded.csv', index=False)

# Sensor 2

df_2 = pd.read_csv(dataset_dir + 'sensor2_clean.csv')

df_2.shape

df_2.head()

df_2 = encode(df_2)
df_2 = add_target(df_2)

df_2.shape

df_2.head()

df_2 = df_2.drop(columns=['ActivityLabel','BotnetName','SensorId'])

df_2.to_csv(dataset_dir + 'sensor2_encoded.csv', index=False)

# Sensor 3

df_3 = pd.read_csv(dataset_dir + 'sensor3_clean.csv')

df_3.shape

df_3.head()

df_3 = encode(df_3)
df_3 = add_target(df_3)

df_3.shape

df_3.head()

df_3 = df_3.drop(columns=['ActivityLabel','BotnetName','SensorId'])

df_3.to_csv(dataset_dir + 'sensor3_encoded.csv', index=False)
