==> Revise_Experiment/Revisi 2/3.SplitDataset.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
from sklearn.model_selection import train_test_split
%matplotlib inline

# tabel setting
pd.set_option('display.max_columns',None)
pd.set_option('display.max_rows',None)
pd.set_option('display.precision', 2)
pd.options.display.float_format = '{:,.2f}'.format

dataset_dir = '../Dataset/'

os.listdir(dataset_dir)

# Split Data Train n Test

def split_data(df):
    train, test = train_test_split(df, test_size=0.25, random_state=42)
    return train, test

# Sensor 1

df_1 = pd.read_csv(dataset_dir + 'sensor1_encoded.csv')

sensor1_train, sensor1_test = split_data(df_1)

sensor1_test.to_csv(dataset_dir + 'sensor1_test.csv', index=False)
sensor1_train.to_csv(dataset_dir + 'sensor1_train.csv', index=False)

## isBotnet DataTrain

isbotnet_1 = sensor1_train.copy()
isbotnet_1.head()

isbotnet_1 = isbotnet_1.drop(columns=['Label', 'isSpam'])

isbotnet_1.head()

isbotnet_1.to_csv(dataset_dir + 'sensor1_isbotnet.csv', index=False)

## isSpam DataTrain

isspam_1 = sensor1_train.copy()
isspam_1 = isspam_1[isspam_1['isBotnet'] == 1]
isspam_1.head()

isspam_1 = isspam_1.drop(columns=['Label', 'isBotnet'])

isspam_1.head()

isspam_1.to_csv(dataset_dir + 'sensor1_isspam.csv', index=False)

# Sensor 2

df_2 = pd.read_csv(dataset_dir + 'sensor2_encoded.csv')

sensor2_train, sensor2_test = split_data(df_2)

sensor2_test.to_csv(dataset_dir + 'sensor2_test.csv', index=False)
sensor2_train.to_csv(dataset_dir + 'sensor2_train.csv', index=False)

## isBotnet DataTrain

isbotnet_2 = sensor2_train.copy()
isbotnet_2.head()

isbotnet_2 = isbotnet_2.drop(columns=['Label', 'isSpam'])

isbotnet_2.head()

isbotnet_2.to_csv(dataset_dir + 'sensor2_isbotnet.csv', index=False)

## isSpam DataTrain

isspam_2 = sensor2_train.copy()
isspam_2 = isspam_2[isspam_2['isBotnet'] == 1]
isspam_2.head()

isspam_2 = isspam_2.drop(columns=['Label', 'isBotnet'])

isspam_2.head()

isspam_2.to_csv(dataset_dir + 'sensor2_isspam.csv', index=False)

# Sensor 3

df_3 = pd.read_csv(dataset_dir + 'sensor3_encoded.csv')

sensor3_train, sensor3_test = split_data(df_3)

sensor3_test.to_csv(dataset_dir + 'sensor3_test.csv', index=False)
sensor3_train.to_csv(dataset_dir + 'sensor3_train.csv', index=False)

## isBotnet DataTrain

isbotnet_3 = sensor3_train.copy()
isbotnet_3.head()

isbotnet_3 = isbotnet_3.drop(columns=['Label', 'isSpam'])

isbotnet_3.head()

isbotnet_3.to_csv(dataset_dir + 'sensor3_isbotnet.csv', index=False)

## isSpam DataTrain

isspam_3 = sensor3_train.copy()
isspam_3 = isspam_3[isspam_3['isBotnet'] == 1]
isspam_3.head()

isspam_3 = isspam_3.drop(columns=['Label', 'isBotnet'])

isspam_3.head()

isspam_3.to_csv(dataset_dir + 'sensor3_isspam.csv', index=False)
